requirements.txt
claude-agent-sdk
anyio
pydantic>=2.0
PyYAML>=6.0
starlette
uvicorn
sse-starlette
httpx
